Refactor linter to share AST index
