+++ Package.swift
@@ -8,21 +8,30 @@ let package = Package(
     ],
     products: [
         .library(name: "DocEngine", targets: ["DocEngine"]),
-        .library(name: "MarkdownLintKit", targets: ["MarkdownLintKit"])
+        .library(name: "MarkdownLintKit", targets: ["MarkdownLintKit"]),
+        .library(name: "MarkdownSupport", targets: ["MarkdownSupport"])
     ],
     dependencies: [
         .package(url: "https://github.com/apple/swift-markdown", from: "0.2.0")
     ],
     targets: [
         .target(
+            name: "MarkdownSupport",
+            dependencies: [
+                .product(name: "Markdown", package: "swift-markdown")
+            ]
+        ),
+        .target(
             name: "DocEngine",
             dependencies: [
+                "MarkdownSupport",
                 .product(name: "Markdown", package: "swift-markdown")
             ]
         ),
         .target(
             name: "MarkdownLintKit",
             dependencies: [
+                "MarkdownSupport",
                 .product(name: "Markdown", package: "swift-markdown")
             ]
         ),
