+++ Sources/MarkdownLintKit/LintContext.swift
@@ -1,4 +1,5 @@
 import Markdown
+import MarkdownSupport
 
 // MARK: - Infrastructure helpers
 public struct LintContext {
