+++ Sources/MarkdownLintKit/MarkdownLinter.swift
@@ -1,4 +1,5 @@
 import Markdown
+import MarkdownSupport
 
 // MARK: - Lint entry
 public enum MarkdownLinter {
@@ -6,7 +7,7 @@ public enum MarkdownLinter {
     public static func lint(_ markdown: String,
                             config: [String: AnyCodable] = [:]) -> [Diagnostic] {
         let document = Document(parsing: markdown)
-        let index = buildIndex(document: document, text: markdown)
+        let index = MarkdownSupport.buildIndex(document: document, text: markdown)
         let context = LintContext(text: markdown, ast: document, index: index, configuration: config)
         var diagnostics: [Diagnostic] = []
         for rule in RuleRegistry.shared.activeRules {
