+++ Sources/MarkdownLintKit/RuleRegistry.swift
@@ -2,17 +2,25 @@
 @MainActor
 public final class RuleRegistry {
     public static let shared: RuleRegistry = RuleRegistry()
-    private var rules: [LintRule] = [
-        RequiredHeadingRule(),
-        UniqueHeadingTitlesRule(),
-        AllowedFencedTagsRule(),
-        NoEmptySectionRule()
-    ]
+    private var rules: [LintRule] = []
 
     private init() {}
 
     public func register(_ rule: LintRule) {
         rules.append(rule)
+    }
+
+    public func registerBuiltInRules() {
+        rules += [
+            RequiredHeadingRule(),
+            UniqueHeadingTitlesRule(),
+            AllowedFencedTagsRule(),
+            NoEmptySectionRule()
+        ]
+    }
+
+    public func reset() {
+        rules.removeAll()
     }
 
     public var activeRules: [LintRule] { rules }
