+++ Sources/MarkdownSupport/Slug.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public enum Slug {
+    public static func slug(_ text: String) -> String {
+        var slug = text.lowercased()
+        slug = slug.replacingOccurrences(of: "[^a-z0-9]+", with: "-", options: .regularExpression)
+        while slug.contains("--") { slug = slug.replacingOccurrences(of: "--", with: "-") }
+        slug = slug.trimmingCharacters(in: CharacterSet(charactersIn: "-"))
+        return slug
+    }
+}
